Misc cleanups to MaybeBounce.
Add comments, rename some variables. Change the way it's called to
remove the "directionName" parameter (which was only there for a debug
writefln anyway).

Cleaned up the code that calls MaybeBounce so that it actually checks
whether there was a bounce. This is a first step toward correctly
handling multiple bounces in a single frame, but we still don't handle
that, as noted in a FIXME in MaybeBounce.